tests(jscad-surface) moved tests suites to tests directory
